fix: replace rank system with dependency-light safe build
The previous RankJourney used heavy framer-motion animations
(animateTransform SMIL, infinite keyframes, nested SVG defs) that
were triggering a runtime crash on render in some browsers. The
new build removes framer-motion entirely from the rank system and
inlines everything as plain CSS + simple inline SVGs.

Changes:
- RankJourney.jsx — now exports CurrentRankBadge + JourneyProgressTrack
  directly with no external imports beyond React + ranks.js
- Deleted BadgeArt.jsx + AchievementCelebration.jsx + old
  JourneyProgressTrack.jsx (the buggy ones)
- Removed the AchievementCelebration toast queueing logic
- All rank badges still render the same 6-tier visual identity
  (BronzeShield / SilverWings / GoldCrystal / DiamondCrystal /
  HeroicCrest / MasterEmblem) as plain SVG, no SMIL animations
- Track still has 16 milestones + runner character
- CurrentRankBadge hero + Next Rank footer preserved

Also added an ErrorBoundary around StudentView (StudentViewErrorBoundary)
that catches render errors and writes them to
localStorage.__spurti_last_error so future debugging is easy.

Build: 768 modules, 76.6 KB CSS / 932.9 KB JS.
